requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# NLP Libraries
nltk>=3.8
spacy>=3.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Utilities
python-multipart>=0.0.6
